Many TkTime improvements and reorganizations
In the future, we will use TkTime to generate a tagged event information
corpus.  There is a new option on the menu, "(not present)" which allows
you to tell TkTime that the right answer was missing (currently, we don't
do anything with this information).

TkTime objects now take a parses_callback parameter which is told when
a segment gets new parses.  TkTime also remembers your selections per
text of the segment.  Thus, editing the context or the main entry will
not change a selection unless the text of that selection is changed.
Possibly, context changes should invalidate this.

TkTime.sched_update supports a force_update argument which it passes to
update_display.  When the context changes, we schedule a forced update
rather than updating immediately (to wait until they stop typing).

TkTimeWithContext is a new widget which bundles two TkTimes together            and treats one as a context for the other (the context for the context
is always "now").  When you type in the context, the first parse is
automatically used as the context.  You can still manually override this
with the RMB menu.

Utility now has a Symbol class, for Lisp-like symbols.  They're like
strings but used differently (think exception classes vs. string
exceptions).  Two symbols of the same name are equal but not equal
to the string of their name.

darcs-hash:20050602023512-78b18-20d79a12936f0a70f9eb080e1315d585ec30912a.gz