expand_ordinal_dow_month() and undeadify (and improve) day of week checking
expand_ordinal_dow_month expands things of the form
"the last Thursday in March" or the "third Wednesday in July 2001".
Day of week checking was actually dead code since it was from when results
were still datetime objects and not PartialTimes.  It is enabled now and
only gives a slight bonus to things that match (the value was tweaked
while watching the output of the test suite.  Also, it will no longer
bail out on dates that fail day of week checking, as this also helped
the scores.  This merits further investigation.

darcs-hash:20050629081506-78b18-4eb19aa9fc3f75a46c4b7fd06fbe7ec754975f4d.gz